Split services and controllers in one file/module
