+++ app/assets/index.html
@@ -29,5 +29,6 @@
 
   <script src="vendor.js"></script>
   <script src="app.js"></script>
+  <script>require('js/application');</script>
 </body>
 </html>
