+++ app/js/application.js
@@ -1,4 +1,11 @@
-"use strict";
+var angular = require('angular');
+
+require('angular-emojify');
+require('angular-tooltips');
+require('angular-route');
+require('angular-local-storage');
+require('angular-moment');
+require('lodash');
 
 angular.module('app', ['emojify', '720kb.tooltips', 'ngRoute', 'LocalStorageModule', 'angularMoment'])
 
@@ -27,6 +34,14 @@ angular.module('app', ['emojify', '720kb.tooltips', 'ngRoute', 'LocalStorageModu
     return _.size(collection);
   }
 })
+
+.service('configManager', require('./services/config_manager'))
+.service('gitLabManager', require('./services/gitlab_manager'))
+.service('favicoService', require('./services/favico'))
+.service('MergeRequestFetcher', require('./services/merge_request_fetcher'))
+
+.controller('DashboardCtrl', require('./controllers/dashboard'))
+.controller('SettingsCtrl', require('./controllers/settings'))
 
 .run(function($rootScope, gitLabManager, $location, $http) {
   $rootScope.titleAddon = '';
